requirements.txt
Flask
kreuzberg
python-magic
requests
Werkzeug
gunicorn
aiofiles